Changed str() to as_string()
Originally I was planning to be able to ask the SCValue object
"what was the original string". However, a combination of not
being able to come up with a real use case for this, and a
realization that to do it right I would need to rewrite a
significant amount of code has made me change my mind.
Since now the user is not going to be able to call str() for
every value type, now it makes more sense for str() to be
called as_string() like everything else.